Organize date utils. Develop top bar to display information received from backend
